requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.76.0
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.3
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
flask-cors==3.0.10
gunicorn==20.1.0
icalendar==5.0.7
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
jsonschema==4.17.3
mangum==0.17.0
mongoengine==0.27.0
motor==3.1.2
passlib==1.7.4
pkgutil_resolve_name==1.3.10
pycparser==2.21
pymongo==4.3.3
PyMySQL==1.0.3
pyrsistent==0.19.3
python-dateutil==2.8.2
python-jose==3.3.0
pytz==2023.3
requests==2.29.0
rsa==4.9
six==1.16.0
twilio==8.2.0
urllib3==1.26.15
Werkzeug==2.3.3
zipp==3.15.0
google-genai
jwt